+++ apps/health_campaign_field_worker_app/lib/pages/home.dart
@@ -8,6 +8,7 @@ import 'package:complaints/complaints.dart';
 import 'package:complaints/router/complaints_router.gm.dart';
 import 'package:connectivity_plus/connectivity_plus.dart';
 import 'package:digit_data_model/data_model.dart';
+import 'package:digit_data_model/models/entities/household_type.dart';
 import 'package:digit_data_model/models/entities/user_action.dart';
 import 'package:digit_dss/data/local_store/no_sql/schema/dashboard_config_schema.dart';
 import 'package:digit_dss/models/entities/dashboard_response_model.dart';
@@ -17,6 +18,7 @@ import 'package:digit_location_tracker/utils/utils.dart';
 import 'package:digit_ui_components/digit_components.dart';
 import 'package:digit_ui_components/utils/component_utils.dart';
 import 'package:drift_db_viewer/drift_db_viewer.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_background_service/flutter_background_service.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -49,6 +51,7 @@ import '../router/app_router.dart';
 import '../utils/debound.dart';
 import '../utils/environment_config.dart';
 import '../utils/i18_key_constants.dart' as i18;
+import '../utils/least_level_boundary_singleton.dart';
 import '../utils/utils.dart';
 import '../widgets/header/back_navigation_help_header.dart';
 import '../widgets/home/home_item_card.dart';
@@ -71,6 +74,7 @@ class _HomePageState extends LocalizedState<HomePage> {
   bool skipProgressBar = false;
   final storage = const FlutterSecureStorage();
   late StreamSubscription<List<ConnectivityResult>> subscription;
+  bool isTriggerLocalisation = true;
 
   @override
   initState() {
@@ -163,8 +167,11 @@ class _HomePageState extends LocalizedState<HomePage> {
                     ),
             ],
           ),
-          footer: PoweredByDigit(
-            version: Constants().version,
+          footer: Padding(
+            padding: const EdgeInsets.only(bottom: spacer2),
+            child: PoweredByDigit(
+              version: Constants().version,
+            ),
           ),
           children: [
             const SizedBox(height: spacer2 * 2),
@@ -330,8 +337,13 @@ class _HomePageState extends LocalizedState<HomePage> {
         child: HomeItemCard(
           icon: Icons.announcement,
           label: i18.home.fileComplaint,
-          onPressed: () =>
-              context.router.push(const ComplaintsInboxWrapperRoute()),
+          onPressed: () {
+            if (isTriggerLocalisation) {
+              triggerLocalization();
+              isTriggerLocalisation = false;
+            }
+            context.router.push(const ComplaintsInboxWrapperRoute());
+          },
         ),
       ),
 
@@ -340,6 +352,10 @@ class _HomePageState extends LocalizedState<HomePage> {
           icon: Icons.bar_chart_sharp,
           label: i18.home.dashboard,
           onPressed: () {
+            if (isTriggerLocalisation) {
+              triggerLocalization();
+              isTriggerLocalisation = false;
+            }
             context.router.push(const UserDashboardRoute());
           },
         ),
@@ -353,6 +369,10 @@ class _HomePageState extends LocalizedState<HomePage> {
           onPressed: () async {
             RegistrationDeliverySingleton()
                 .setHouseholdType(HouseholdType.family);
+            if (isTriggerLocalisation) {
+              triggerLocalization();
+              isTriggerLocalisation = false;
+            }
             await context.router.push(const RegistrationDeliveryWrapperRoute());
           },
         ),
@@ -365,6 +385,10 @@ class _HomePageState extends LocalizedState<HomePage> {
           onPressed: () async {
             RegistrationDeliverySingleton()
                 .setHouseholdType(HouseholdType.community);
+            if (isTriggerLocalisation) {
+              triggerLocalization();
+              isTriggerLocalisation = false;
+            }
             await context.router.push(const RegistrationDeliveryWrapperRoute());
           },
         ),
@@ -378,6 +402,10 @@ class _HomePageState extends LocalizedState<HomePage> {
           customIcon: Constants.closedHouseholdSvg,
           label: i18.home.closedHouseHoldLabel,
           onPressed: () {
+            if (isTriggerLocalisation) {
+              triggerLocalization();
+              isTriggerLocalisation = false;
+            }
             context.router.push(const ClosedHouseholdWrapperRoute());
           },
         ),
@@ -388,16 +416,11 @@ class _HomePageState extends LocalizedState<HomePage> {
           icon: Icons.store_mall_directory,
           label: i18.home.manageStockLabel,
           onPressed: () {
-            context.read<AppInitializationBloc>().state.maybeWhen(
-                  orElse: () {},
-                  initialized: (
-                    AppConfiguration appConfiguration,
-                    _,
-                    __,
-                  ) {
-                    context.router.push(ManageStocksRoute());
-                  },
-                );
+            if (isTriggerLocalisation) {
+              triggerLocalization();
+              isTriggerLocalisation = false;
+            }
+            context.router.push(ManageStocksRoute());
           },
         ),
       ),
@@ -407,6 +430,10 @@ class _HomePageState extends LocalizedState<HomePage> {
           icon: Icons.menu_book,
           label: i18.home.stockReconciliationLabel,
           onPressed: () {
+            if (isTriggerLocalisation) {
+              triggerLocalization();
+              isTriggerLocalisation = false;
+            }
             context.router.push(StockReconciliationRoute());
           },
         ),
@@ -415,9 +442,17 @@ class _HomePageState extends LocalizedState<HomePage> {
         child: HomeItemCard(
           enableCustomIcon: true,
           customIcon: mySurveyFormSvg,
+          iconPadding: const EdgeInsets.all(spacer1),
           icon: Icons.checklist,
+          customIconSize: spacer8,
           label: i18.home.mySurveyForm,
-          onPressed: () => context.router.push(SurveyFormWrapperRoute()),
+          onPressed: () {
+            if (isTriggerLocalisation) {
+              triggerLocalization();
+              isTriggerLocalisation = false;
+            }
+            context.router.push(SurveyFormWrapperRoute());
+          },
         ),
       ),
 
@@ -452,16 +487,11 @@ class _HomePageState extends LocalizedState<HomePage> {
           icon: Icons.supervised_user_circle_rounded,
           label: i18.home.beneficiaryReferralLabel,
           onPressed: () async {
-            context.read<AppInitializationBloc>().state.maybeWhen(
-                  orElse: () {},
-                  initialized: (
-                    AppConfiguration appConfiguration,
-                    _,
-                    __,
-                  ) {
-                    context.router.push(SearchReferralReconciliationsRoute());
-                  },
-                );
+            if (isTriggerLocalisation) {
+              triggerLocalization();
+              isTriggerLocalisation = false;
+            }
+            context.router.push(SearchReferralReconciliationsRoute());
           },
         ),
       ),
@@ -470,6 +500,10 @@ class _HomePageState extends LocalizedState<HomePage> {
           icon: Icons.announcement,
           label: i18.home.viewReportsLabel,
           onPressed: () {
+            if (isTriggerLocalisation) {
+              triggerLocalization();
+              isTriggerLocalisation = false;
+            }
             context.router.push(InventoryReportSelectionRoute());
           },
         ),
@@ -480,6 +514,11 @@ class _HomePageState extends LocalizedState<HomePage> {
           icon: Icons.fingerprint_outlined,
           label: i18.home.manageAttendanceLabel,
           onPressed: () {
+            if (isTriggerLocalisation) {
+              triggerLocalization();
+              isTriggerLocalisation = false;
+            }
+            ;
             context.router.push(const ManageAttendanceRoute());
           },
         ),
@@ -504,6 +543,11 @@ class _HomePageState extends LocalizedState<HomePage> {
           icon: Icons.bar_chart_sharp,
           label: i18.home.dashboard,
           onPressed: () {
+            if (isTriggerLocalisation) {
+              triggerLocalization();
+              isTriggerLocalisation = false;
+            }
+            ;
             context.router.push(const UserDashboardRoute());
           },
         ),
@@ -568,8 +612,7 @@ class _HomePageState extends LocalizedState<HomePage> {
                 .map((e) => e.displayName)
                 .toList()
                 .contains(element) ||
-            element == i18.home.db ||
-            element == i18.home.transitPostLabel)
+            element == i18.home.db)
         .toList();
 
     final showcaseKeys = filteredLabels
@@ -577,8 +620,8 @@ class _HomePageState extends LocalizedState<HomePage> {
         .map((label) => homeItemsShowcaseMap[label]!)
         .toList();
 
-    if (!context.selectedProject.name.contains('IRS')) {
-      filteredLabels.remove(i18.home.dashboard);
+    if (envConfig.variables.envType == EnvType.demo && kReleaseMode) {
+      filteredLabels.remove(i18.home.db);
     }
 
     final List<Widget> widgetList =
@@ -671,6 +714,32 @@ class _HomePageState extends LocalizedState<HomePage> {
           );
     }
   }
+
+  void triggerLocalization() {
+    context.read<AppInitializationBloc>().state.maybeWhen(
+          orElse: () {},
+          initialized: (
+            AppConfiguration appConfiguration,
+            _,
+            __,
+          ) {
+            final appConfig = appConfiguration;
+            final localizationModulesList = appConfiguration.backendInterface;
+            final selectedLocale = AppSharedPreferences().getSelectedLocale;
+            LocalizationParams()
+                .setCode(LeastLevelBoundarySingleton().boundary);
+            context
+                .read<LocalizationBloc>()
+                .add(LocalizationEvent.onLoadLocalization(
+                  module:
+                      "${localizationModulesList?.interfaces.where((element) => element.type == Modules.localizationModule).map((e) => e.name.toString()).join(',')}",
+                  tenantId: appConfig.tenantId ?? "default",
+                  locale: selectedLocale!,
+                  path: Constants.localizationApiPath,
+                ));
+          },
+        );
+  }
 }
 
 // Function to set initial Data required for the packages to run
@@ -680,8 +749,10 @@ void setPackagesSingleton(BuildContext context) {
       initialized: (
         AppConfiguration appConfiguration,
         List<ServiceRegistry> serviceRegistry,
-        DashboardConfigSchema? dashboardConfigSchema,
+        List<DashboardConfigSchema?>? dashboardConfigSchema,
       ) {
+        final filteredDashboardConfig = filterDashboardConfig(
+            dashboardConfigSchema ?? [], context.projectTypeCode ?? "");
         loadLocalization(context, appConfiguration);
         // INFO : Need to add singleton of package Here
         TransitPostSingleton().setInitialData(
@@ -820,7 +891,7 @@ void setPackagesSingleton(BuildContext context) {
         DashboardSingleton().setInitialData(
             projectId: context.projectId,
             tenantId: envConfig.variables.tenantId,
-            dashboardConfig: dashboardConfigSchema,
+            dashboardConfig: filteredDashboardConfig.firstOrNull,
             appVersion: Constants().version,
             selectedProject: context.selectedProject,
             actionPath: Constants.getEndPoint(