+++ apps/health_campaign_field_worker_app/lib/utils/constants.dart
@@ -74,7 +74,7 @@ class Constants {
           OpLogSchema,
           ProjectTypeListCycleSchema,
           RowVersionListSchema,
-          DashboardConfigSchemaSchema,
+          DashboardConfigSchemaListSchema,
           DashboardResponseSchema,
         ],
         name: 'HCM',
@@ -94,6 +94,8 @@ class Constants {
   static const String healthFacilitySurveyFormPrefix = 'HF_RF';
 
   static const String boundaryLocalizationPath = 'rainmaker-boundary-admin';
+
+  static const String dashboardAnalyticsPath = '/dashboard-analytics/dashboard/getChartV2';
 
   static RegExp mobileNumberRegExp =
       RegExp(r'^(?=.{10}$)[+]*[(]{0,1}[0-9]{1,4}[)]{0,1}[-\s\./0-9]*$');