+++ packages/registration_delivery/lib/data/repositories/local/individual_global_search.dart
@@ -183,7 +183,8 @@ class IndividualGlobalSearchRepository extends LocalRepository {
                 .equalsExp(sql.individual.clientReferenceId))
       ])
         ..where(buildAnd([
-          if (params.householdType == HouseholdType.community && params.householdClientReferenceId==null)
+          if (params.householdType == HouseholdType.community &&
+              params.householdClientReferenceId == null)
             sql.householdMember.isHeadOfHousehold.equals(true),
           sql.address.relatedClientReferenceId.isNotNull(),
           sql.individual.clientReferenceId.isNotNull(),
@@ -282,9 +283,10 @@ class IndividualGlobalSearchRepository extends LocalRepository {
           leftOuterJoin(
               sql.projectBeneficiary,
               sql.projectBeneficiary.beneficiaryClientReferenceId
-                  .equalsExp(sql.household.clientReferenceId))
+                  .equalsExp(sql.individual.clientReferenceId))
         ]).where(buildAnd([
-          if (params.householdType == HouseholdType.community && params.householdClientReferenceId==null)
+          if (params.householdType == HouseholdType.community &&
+              params.householdClientReferenceId == null)
             sql.householdMember.isHeadOfHousehold.equals(true),
           sql.household.householdType.equalsValue(params.householdType)
         ]));