+++ apps/health_campaign_field_worker_app/lib/blocs/app_initialization/app_initialization.dart
@@ -316,6 +316,7 @@ class MdmsConfig {
   final List<ServiceRegistry> serviceRegistryList;
   final List<DashboardConfigSchema?>? dashboardConfigSchema;
 
+
   const MdmsConfig(
       {required this.appConfigs,
       required this.serviceRegistryList,